cloud/deployment/requirement.txt
pandas
numpy
plotly
matplotlib
seaborn
Pillow
azure-storage-blob
scikit-learn
opencv-python
opencv-python-headless
opencv-contrib-python
torch
torchvision
torchaudio
jupyterlab-lsp
python-lsp-server
jupyter
notebook
python-dotenv
chardet
tqdm
torchserve
sagemaker
boto3